// frontend/src/app/core/shared/common/enum/purchase-order.enum.ts
export enum PurchaseOrderEnum {
    source = 'source',
    po_no = 'po_no',
    po_date = 'po_date',
    vendor = 'vendor',
    poc = 'poc',
    vendor_quotation_no = 'vendor_quotation_no',
    quotation_date = 'quotation_date',
    delivery_terms = 'delivery_terms',
    delivery_preference = 'delivery_preference',
    payment_terms = 'payment_terms',
    bill_to_name = 'bill_to_name',
    bill_to_address1 = 'bill_to_address1',
    bill_to_address2 = 'bill_to_address2',
    bill_to_tax_no = 'bill_to_tax_no',
    bill_to_country = 'bill_to_country',
    bill_to_city = 'bill_to_city',
    bill_to_zip_code = 'bill_to_zip_code',
    bill_to_state = 'bill_to_state',
    ship_to_name = 'ship_to_name',
    ship_to_address1 = 'ship_to_address1',
    ship_to_address2 = 'ship_to_address2',
    ship_to_tax_no = 'ship_to_tax_no',
    ship_to_country = 'ship_to_country',
    ship_to_city = 'ship_to_city',
    ship_to_zipcode = 'ship_to_zipcode',
    ship_to_state = 'ship_to_state',
    sub_total = 'sub_total',
    discount = 'discount',
    tax_amount = 'tax_amount',
    total_amount = 'total_amount',
    payment_type = 'payment_type',
    note = 'note',
    terms_conditions = 'terms_conditions',
    purchase_order_items = 'purchase_order_items',
    poc_name = "poc_name",
    poc_phone = "poc_phone",
    poc_email = "poc_email"
}

export enum PurchaseOrderItemEnum {
  asset_code = 'asset_code',
  asset_name = 'asset_name',
  asset_description = 'asset_description',
  asset_mfr = 'mfr',
  delivery_date = 'delivery_date',
  qty = 'qty',
  uom = 'uom',
  unit_price = 'unit_price',
  discount = 'discount',
  discount_amount = 'discount_amount',
  tax = 'tax',
  tax_amount = 'tax_amount',
  total_amount = 'total_amount',
}